+++ crates/acts/src/cache/mod.rs
@@ -1,8 +1,8 @@
 #[allow(clippy::module_inception)]
 mod cache;
-mod writer;
 mod store;
 #[cfg(test)]
 mod tests;
+mod writer;
 
 pub use cache::Cache;